requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.1.6
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
pillow
django-cleanup